requirements.txt
six
txaio
pyasn1
autobahn
crossbar
trollius
rsb-python
